🎨 Move contact component into cv page
